123/requirements.txt
gradio>=4.0.0
Pillow>=10.0.0
